+++ spec/lib/dawn/codesake_knowledgebase_spec.rb
@@ -1049,4 +1049,9 @@ describe "The Codesake Dawn knowledge base" do
     sc.should_not   be_nil
     sc.class.should == Dawn::Kb::OSVDB_121701
   end
+  it "must have test for CVE-2015-4020" do
+    sc = kb.find("CVE-2015-4020")
+    sc.should_not   be_nil
+    sc.class.should == Dawn::Kb::CVE_2015_4020
+  end
 end